계산식 변경

+++ Code/background.py
@@ -13,8 +13,8 @@ class Background:
         self.x = 0
 
     def update(self):
-        self.x = int((self.x + (RUN_SPEED_PPS * game_framework.frame_time)) % 980)
+        self.x = (self.x + RUN_SPEED_PPS * game_framework.frame_time) % 980
 
     def draw(self):
-        self.image.clip_draw(self.x, 0, 980 - self.x, 560, 490 - self.x / 2.0, 280)
-        self.image.clip_draw(0, 0, self.x, 560, 980 - self.x / 2.0, 280)
+        self.image.clip_draw(int(self.x), 0, 980 - int(self.x), 560, 490 - int(self.x) / 2.0, 280)
+        self.image.clip_draw(0, 0, int(self.x), 560, 980 - int(self.x) / 2.0, 280)
